requirements.txt
groq==0.19.0
langchain_community==0.3.19
langchain_core==0.3.45
langchain_groq==0.2.5
langgraph==0.3.10
pymongo==4.11.2
python-dotenv==1.0.1
streamlit==1.43.2
typing_extensions==4.12.2
